int->size_t

+++ libraries/OpenravePlugins/OpenraveYarpPluginLoader/OpenraveYarpPluginLoader.cpp
@@ -46,14 +46,14 @@ public:
     virtual ~OpenraveYarpPluginLoader()
     {
         //-- Note that we start on element 1, first elem was not via new!!
-        for(int i=1;i<argv.size();i++)
+        for(size_t i=1;i<argv.size();i++)
         {
             //CD_DEBUG("Deleting [%s]\n",argv[i]);
             delete argv[i];
             argv[i] = 0;
         }
 
-        for(int i=0;i<yarpPlugins.size();i++)
+        for(size_t i=0;i<yarpPlugins.size();i++)
         {
             yarpPlugins[i]->close();
             delete yarpPlugins[i];
